--- src/components/miscselector/miscselector.module.css
@@ -1,103 +1,0 @@
-.card {
-    box-sizing: border-box;
-    background: rgb(213, 217, 206);
-    background: linear-gradient(0deg, rgba(213, 217, 206, 1) 0%, rgba(238, 239, 231, 1) 75%, rgba(255, 255, 255, 1) 100%);
-    border: white solid 0.3rem;
-    margin: 1rem;
-    padding: 2rem 0;
-    border-radius: 1rem;
-    display: flex;
-    justify-content: center;
-    color: #5f5f5f;
-    box-shadow: 0 0.4rem 0.4rem rgb(55, 55, 55);
-}
-
-.slider {
-    -webkit-appearance: none;
-    border-radius: 0.3rem;
-    height: 1rem;
-    width: 8.8rem;
-    background: white;
-    margin: 0.13rem 0.5rem;
-    transform: translateY(0.11rem);
-}
-
-.slider::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    appearance: none;
-    width: 1.5rem;
-    height: 1.5rem;
-    border-radius: 30%;
-    cursor: pointer;
-    border: 2px white solid;
-    background-color: #5f5f5f;
-    transition: ease-in-out 300ms transform;
-}
-
-.slider::-webkit-slider-thumb:hover {
-    transform: scale(120%);
-}
-
-.card span {
-    font-size: 1.1rem;
-}
-
-.wardenLvl {
-    width: 2rem;
-    height: 1.5rem;
-    border: none;
-    border-radius: 0.2rem;
-    margin: 0 0.3rem;
-}
-
-.defLvl {
-    width: 3.5rem;
-    height: 1.5rem;
-    border: none;
-    border-radius: 0.2rem;
-    margin: 0 0.3rem;
-}
-
-.warden {
-    margin-top: 0.4rem;
-    display: block;
-}
-
-.level {
-    margin-top: 0.7rem;
-    display: block;
-}
-
-input {
-    font-family: supercell;
-    font-size: 1.1rem;
-    text-align: center;
-}
-
-.select {
-    margin: 0 0 1rem 0;
-    z-index: 5;
-}
-
-.calcBtn {
-    width: 13rem;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 1.4rem;
-    display: block;
-    cursor: pointer;
-    transition: ease-in-out 300ms transform;
-}
-
-.calcBtn:hover {
-    transform: scale(103%);
-}
-
-.currentCC {
-    width: 1.2rem;
-    display: inline-block;
-}
-
-.sliders {
-    margin: 0.5rem 0;
-}
