--- src/components/navbar/navbar.module.css
@@ -1,18 +1,0 @@
-nav {
-    height: 5rem;
-    background: linear-gradient(0deg, rgba(54, 54, 54, 1) 30%, rgba(76, 76, 74, 1) 100%);
-    border-bottom: 0.5rem solid rgb(50, 50, 50);
-    border-top: 0.5rem solid rgb(83, 82, 82);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.content {
-    text-align: center;
-}
-
-.navText {
-    width: 16rem;
-    cursor: pointer;
-}
