+++ src/components/spellimg/spellimg.module.css
@@ -1,6 +1,6 @@
 .spellImg {
-    width: 8rem;
-    border-radius: 1rem;
+    width: 4.5rem;
+    border-radius: 0.5rem;
     border: 0.15rem white solid;
 }
